+++ gff3toddbj/transforms.py
@@ -257,11 +257,21 @@ def _join_features(record: SeqRecord, joinables: Optional[Tuple[str, ...]]) -> S
         return result
 
 
-    def _runner(features: List[SeqFeature]) -> List[SeqFeature]:
+    def _runner(features: List[SeqFeature], parent: Optional[str]) -> List[SeqFeature]:
         """Scan features and apply _join """
-        # `groups_or_features` has either `SeqFeature` or a group == (type, product) as its key.
-        # An item with a `SeqFeature` key has dummy while an item with the tuple key has
-        # List[SeqFeature] as its value.
+
+        # Don't join features if the parent is None or "gene"
+        if (parent is None) or (parent == "gene"):
+            for f in features:
+                if hasattr(f, "sub_features") and f.sub_features:
+                    f.sub_features = _runner(f.sub_features, f.type)
+            return features
+
+        # Otherwise, group features that are joined.
+        # `groups_or_features` is a bag both to-be-joined and not-joined features.
+        # Type of `groups_or_features` key is either `SeqFeature` or a group == (type, product).
+        # Values for `SeqFeature` keys are dummy, while the type of values for the tuple keys are
+        # List[SeqFeature] that are to be joined.
         groups_or_features = collections.defaultdict(list)
         for f in features:
             if f.type not in joinables_:
@@ -279,7 +289,7 @@ def _join_features(record: SeqRecord, joinables: Optional[Tuple[str, ...]]) -> S
         for group_or_f, fs in groups_or_features.items():
             if isinstance(group_or_f, SeqFeature):
                 res.append(group_or_f)
-            elif len(fs) == 1:
+            elif len(fs) == 1 or parent == "gene":
                 res.extend(fs)
             else:
                 if group_or_f not in seen:
@@ -287,17 +297,17 @@ def _join_features(record: SeqRecord, joinables: Optional[Tuple[str, ...]]) -> S
                     joined_feature = _join(fs)
                     res.append(joined_feature)
 
-        # join after upper levels
+        # join sublevels after the current level
         for f in res:
             if hasattr(f, "sub_features") and f.sub_features:
-                f.sub_features = _runner(f.sub_features)
+                f.sub_features = _runner(f.sub_features, f.type)
 
         return res
 
-    # [NOTE] Do not merge features at the top-level
-    for feature in record.features:
-        if hasattr(feature, "sub_features"):
-            feature.sub_features = _runner(feature.sub_features)
+
+    # --- body of _join_features ---
+    # [NOTE] Do not join features at the top-level (So set the parent as None)
+    record.features = _runner(record.features, parent=None)
 
     return record