+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Current
 * Add a non-SO wild items to the default config for feature and qualifier names
-* Fix: Let feature-wise /transl_table precedes the global setting.
+* Fix: Respect feature-wise /transl_table value than the globally-set one.
 
 ## v0.2.4 (2021-10-04)
 * Expand the default config for feature and qualifier names