+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Current
+## v0.3.0 (2021-10-12)
 * Add `genbank-to-ddbj` for Genbank → DDBJ conversion
 * Add `compare-ddbj`s for comparing two DDBJ annotations for evaluation
 * Add a non-SO wild items to the default config for feature and qualifier names